IDP-774 Custom object injection for scripting contexts: Attributes
https://issues.shibboleth.net/jira/browse/IDP-774

This change adds a customObject to the Scripted AttributeResolver, DataConnector, Matcher, and PolicyRule.
It extends the syntax for the configuration of these (in the afp namespace only for the last two) to allow an
attribute customObjectRef to supplied the name of the bean, with "shibboleth.CustomScriptObject" being
the default.  

The (no system) global.xml gains a bean of this name (otherwise the scripted parsers would fail to find the
bean and sadness would ensue.

Tests for the objects and for the parser.