IDP-694 Deprecate the use of workContext in the Scripted Attribute Plugins
https://issues.shibboleth.net/jira/browse/IDP-694

As a historic accident this context got added to the scripting context for scripted attributes and data connectors.  It can provide no value to a script.  This checkin introduces a delegating object which does the work if needed, but which also logs its use as an error.  Only the specific methods that an AttributeResovlerWorkContext provides are logged, the rest are available, but probably even more pointless.