IDP-1047 Move some constructor beans to setProperty/Initialize
https://issues.shibboleth.net/jira/browse/IDP-1047

In some cases where there is a risk that the bean is being used in production
(despite the impl status) we have left the legacy constructors in place as deprecated
(and warning as such).  

Where, in addition, the class has become initializable the constructor also does the
initialize, swallowinging the Initialization exception and turning it into the
ConstraintException that would have occurred.  Bleargp.