IDP-680 Introduce two new Dependency Types to the schema
https://issues.shibboleth.net/jira/browse/IDP-680

Starting to fix the Dependency mess from the end with the schema change.

The precise syntax can be changed for now we have

    <InputDataConnector ref="dep2" allAttributes="true"/>
    <InputDataConnector ref="dep2" attributeNames="#{ {'Hugh', 'Pugh', 'Barney', 'Mcgrew', 'Cuthbert', 'Dibble', 'Grub'} }"/>
    <InputAttributeDefinition ref="dep3"/>

(noting the use of a spring expressing in the list)

Add parsers and tests, warn on the old style <Dependency/> also update the schema for all the "rolled up" plugins.