IDP-701 Use attribute authority for CAS proxy.
Let the presence of an AttributeAuthorityDescriptor that advertises support
for https://www.apereo.org/cas/protocol indicate CAS proxy authorization.
In the future any signing certificates will be added as explicit trust
material for use in proxy callback validation, which would be analogous to
the mechanics of TLS trust for back-channel attribute queries.