IDP-1344 Implement delegated methods from BaseContext
https://issues.shibboleth.net/jira/browse/IDP-1344

The DelegatedWorkContext (which is there purely to stop people using it)
was trying to get away without implementing the methods that BaseContext
implements.  This is doomed to failure (if anyone had been calling these
methods nothing goo would have happened).

As a side effect we were also calling a couple of deprecated methods which
triggered warning in logs.

Fix this all by implementing all the (non protected) methods.