Checkstyle found bug in taglibs
We were comparing an enum with a string, which checkstyle complained about.
We should have been comapring string with strine, so do this.

Note that this code looks like a straight V2 port (by virtue of having
a private static final long serialVersionUID ....)  So this is probably
a very old bug. Or not a bug at all, but it is cleaner like this.