IDP-1455 Allow data Connectors to export attributes
https://issues.shibboleth.net/jira/browse/IDP-1455

Most of the work is just colelcting and storing the control
values in various beans.

All the heavy lifting is done during attribute resolution where we will
provoke resolution of exporting data connectors and then inject exported
attributes after attribute resolution is done.

Much of the change to the resolver is to refactor out code which is now
common to recording attribute resolution or exporting attributes from
data connectors.