IDP-1450 Review feedback
https://issues.shibboleth.net/jira/browse/IDP-1450

We want to match the type of the setters and the getters so just
changing just the setters to Boolean (from boolean) won't work.

Adding a new setter with the same name and a Boolean also doesn't work
because Spring chooses the boolean setter - even when given a Boolean
value, which blows up with nulls, the avoidance of which is the point
of the entire exercise.