OSJ-272: Move HttpClient TLS TrustEngine use to a X509TrustManager impl
The type of concrete exception that bubbles up from CAS
HttpClientProxyValidator changes slightly on cert
validation failure. With the new TrustManger approach for HttpClient
TLS, we now (I think properly) get an SSLException wrapping a
CertificateException, and this case is handled specifically in the CAS
component.  So the control data in the test needed to change
accordingly.