IDP-1668 Add some basic defense against badly formed paths
https://issues.shibboleth.net/jira/browse/IDP-1668

The concern is that the plugin could inject a path which is outside
where we should be operating ("../../etc/passwd") or some such.

To that end we always canonicalize the source (distribution dir) and
target (idpHome) and then we also check (via canonicalized paths)
that every from file comes from the distribution and every to file
is inside idpHome.