IDP-1683 Substantial rewrite of Plugin Installer
https://issues.shibboleth.net/jira/browse/IDP-1683

Add transactional support for update and install.

Recast the lifetimes of the ClasspathLoaders to allow
Module enabling to work with an active ClasspathLoader.

Do the Module enable/disable thing (transactionally).

Move the Module probing to inside the transactional scope
and after the uninstall part of an upgrade, also give the
Service loader a ClasspathLoader which contains the plugins.