IDP-1683 Plugins not all go into a common folder in dist
We now install all plugins into dist\plugin-webapp.
The installer knows to look there and the runclass scripts do too
(unless they are told not to by the "plugin" command).

Fix a hand full of bugs flished by this and further testing
(notably some classpath issues when emitting the license).