IDP-1660 Sort multi-value attributes before hashing them in for consent
https://issues.shibboleth.net/jira/browse/IDP-1660

Added tests for this.  One of the test attributes get three values
and you get to chose the order. Defaults to sorted for fuiture ease.

Add a new test which explicitly changes the value order before
adding them to the previous context.  Currently fails.

Update all the inline hashes in the tests now that the template
has changed