JPAR-182 Signature checking on WAR contents
https://issues.shibboleth.net/jira/browse/JPAR-182

The needs of the war checking and the plugin handler have diverged
enough that it makes sense to separate the gpg interface classes.

The plugin one goes back to what it was.

Dependency checking uses a substantially cut down clone which is
renamed "KeyRing" and all uses of the word truststore have been exised.