IDP-1866 - Module handling does not work when idpHome is classpath
https://shibboleth.atlassian.net/browse/IDP-1866

On Windows, the (String) parameter to Path#startsWith(String)
must be something which Java things is a valid Path.

Despite "classpath:foo" being a valid Windows file name, Java
disagrees.  the netresult is that path.startWith("classpath:")
throws an exception.

Fix this by casting the Path to a string and doing
String#StartsWith(String)