IDP-1669 Accessibility on IdP pages
https://shibboleth.atlassian.net/browse/IDP-1669

Always freshen up edit-webapp/css and edit-webapp-images.
Rework the fix to use the same code that we use for
updating other directrories (if there is a file with that
name then dont copy).

Also update the copy from distribution to webapp to
preserve the modified date.  This was the date in webapp ends
up being when the distribution was created, not when
the install happened.