IDP-1907 - Enable Spring MVC's CORS-handling features
https://shibboleth.atlassian.net/browse/IDP-1907

Inject shibboleth.CorsConfiguration bean to the FlowHandlerMapping. If the
bean is null / non-existing, the behaviour should remain as before injecting
the bean. If it's set, it's supposed to be a map of CorsConfigurations.

For instance:

    <util:map id="shibboleth.CorsConfigurations" value-type="org.springframework.web.cors.CorsConfiguration">
        <entry key="/oidc/userinfo">
            <bean class="org.springframework.web.cors.CorsConfiguration"
                p:allowedOrigins="http://localhost:8080,http://localhost:8081"
                p:allowedMethods="GET,POST"
                p:allowedHeaders="Authorization"
                p:maxAge="1800" />
        </entry>
    </util:map>

The keys in the map correspond to the locations under the <context>/profile URL space.