Try to cope with Windows Defender
WD has a habit of pinning files and not letting go when the file is deleted.
Because of NTFS default semantics (it does not unlink until the file is
deleted unless you tell it to) this leaves the directory non-empty so the
directory delete then fails.

This is just a test so I'll throw in a "wait 10ms and try again" to bodge this.