IDP-2099 Change the idp-conf tests to rely on file system for idp.home
https://shibboleth.atlassian.net/browse/IDP-2099

Premininary checkin.  Includes test code to populate the
idpHome from
 idp-conf\src\main\resources\
 idp-conf\src\test\resources\
 idp-conf-impl\src\main\resources\

Which is needed until such time as everything is a module.

HOWEVER this requires local editing to point to the on disk directory so that code
is commented out and idp.home and idp.webflows still point to the classpath.